==> code/1.data_preprocessing.ipynb <==
import pandas as pd
import logging
import numpy as np
from sklearn.preprocessing import LabelEncoder
logging.basicConfig(filename="1.Data_proprocessing.log", level=logging.DEBUG, format='%(asctime)s %(message)s',
                    datefmt="%Y-%m-%d %H:%M:%S",filemode="a") 

URL="https://archive.ics.uci.edu/ml/machine-learning-databases/thyroid-disease/thyroid0387.data"
try:
    logging.info("Data preprocessing Started--------")
    try:
        df =  pd.read_csv(URL)
        logging.info("Sucessfully load the dataset ")
    except Exception as ex:
        logging.info(f"error raise when s-----: {ex}")
    try:
        logging.info("Analysis of dataset...........................")
        logging.info(f"Shape of the dataset:{df.shape}")
        logging.info(f"Stastics Analysis :{df.info}")
        logging.info("Data preprocessing started.........")
    except Exception as e:
        logging.info(f"The error raise when the Analysis of dataset is: {e}")
    try:
        feature_cols = ["age","sex","thyroxine", "query_thyroxine", "medication", "sick", "pregnant", "surgery",
                "I131_treatment", "query_hypothyroid", "query_hyperthyroid", "lithium", "goitre", "tumor",
                "hypopituitary", "psych", "TSH_measured", "TSH", "T3_measured", "T3", "TT4_measured", "TT4",
                "T4U_measured", "T4U", "FTI_measured", "FTI", "TBG_measured", "TBG","other","target"]
        logging.info("Sucessfully Rename the Columns")
    except Exception as e:
        logging.info(f"The error raise when the Rename the Columns: {e}")
    try:
        df.columns = feature_cols
        logging.info("Sucessfully Rename the columns Sucessfully")
    except Exception as ex:
        logging.warning(f"Rename cannot be done due to------: {ex}")
    try:
        df["outcome"] = df["target"].str[0]
        df.drop(columns="target", inplace=True)
        logging.info("Sucessfully Extrated information from the target variable.")
    except Exception as e:
        logging.info(f"Error is : {e}")
        logging.info("Replace the alphabet into None variable")
    try:
        ls= ['S', 'F', 'A', 'R', 'I', 'M', 'N', 'G', 'K', 'L', 'Q', 'J',
       'C', 'O', 'H', 'D', 'P', 'B', 'E']
        df['outcome'].replace(ls, "yes", inplace=True)
        logging.info("Sucessfully Replaced all thyroidal diseases into one variable")
    except Exception as e:
        logging.info(f"The error Replaced all thyroidal diseases into one variable is : {e} ")
    try:
        df.outcome.replace({"-":0, "yes":1}, inplace=True)
        logging.info("Sucessfully Classified the target features into 0 and 1.")
    except Exception as e:
        logging.info(f"The error in One hot encoded is : {e}")
    try:
        df.replace("?",np.nan, inplace=True)
        logging.info("Sucessfully Replaced all ? into Null values.")
    except Exception as e:
        logging.info(f"The error in Replacing the value is : {e}")
    try:
        df.drop(columns=["TBG", "T3"], inplace=True)
        logging.info("Sucessfully Droped column TBG and T3 for high missing values.")
    except Exception as e:
        logging.info(f"The error when  Droped column TBG and T3 is : {e}")
    try:
        df.sex.fillna("unknown", inplace=True)
        logging.info("sucessfully imputing the unknown where male and female not mentioned")
    except Exception as e:
        logging.info(f"The error raise when imputing the unknown into sex column is : {e}")
    try: 
        index_age = df[df["age"]>100].index
        df.drop(index_age, inplace=True)
        logging.info("Sucessfully Droped outliers from age feature.")
    except Exception as e:
        logging.info(f"error raise when Droped outliers from age feature is : {e}")
    try:
        df.TSH = pd.to_numeric(df.TSH)
        df.TT4 = pd.to_numeric(df.TT4)
        df.T4U = pd.to_numeric(df.T4U)
        df.FTI = pd.to_numeric(df.FTI)
        logging.info("Sucessfully convert string into Numeric")
    except Exception as e:
        logging.info(f"error raise when Droped outliers from age feature is : {e}")
    try:
        ls=["TSH","FTI","TT4","T4U"]
        for i in ls:
            df[i]=df[i].fillna(df[i].mean())
        logging.info(f"Sucessfully fill the NAN values ")
    except Exception as e:
        logging.info(f"The error raise when fill the NAN values is : {e}")
    le=LabelEncoder()
    X=df.iloc[:,:28]
    try:
        for i in X.columns:
            X[i]=le.fit_transform(X[i])
        logging.info("Sucessfully Encoded the categorical features")
    except Exception as e:
        print(f"The error raise in Encoded the categorical features {e}")
    try:
        logging.info(f"To view the preprocessed dataset: input : {X}")
    except Exception as e:
        logging.info(f"The error raise when view the preprocessed dataset:{e}")
    try:
        logging.info("Saving the processed data ........")
        X.to_csv("data_preprocessed.csv", index=False)
        logging.info("Successfully saved the data in a csv file.")
    except Exception as e:
        logging.info(f"The erro Raise when saved the data in a csv file{e}")
    logging.info("Sucessfully Data preprocessed----------------------------")
except Exception as e:
    loggging.info(f"Data preprocessing not completed {e}")
