==> code/2.train_test_split.ipynb <==
import logging
import pandas as pd
from sklearn.model_selection import train_test_split
logging.basicConfig(filename="2.train_test_split.log", level=logging.INFO, format='%(asctime)s %(message)s',
                    datefmt="%Y-%m-%d %H:%M:%S")
try:
    logging.info("Start the model selection ")
    try:
        df =  pd.read_csv("data_preprocessed.csv")
        logging.info("Sucessfully load the dataset ")
    except Exception as ex:
        logging.info(f"error raise when load the dataset: {ex}")
    try:
        X=df.iloc[:,:27]
        Y=df.iloc[:,27:]
        logging.info("Sucessfully Split the dataset in X and Y format ")
    except Exception as ex:
        logging.info(f"error raise when Split the dataset in X and Y format: {ex}")
        
    try:
        x_train, x_test, y_train, y_test = train_test_split(X,Y,stratify=Y, test_size=0.2)
        logging.info("Sucessfully Split the dataset in train,test and validation format ")
    except Exception as ex:
        logging.info("Sucessfully Split the dataset in train,test and validation format ")
    try:
        x_train.to_csv("x_train.csv", index=False)
        y_train.to_csv("y_train.csv", index=False)
        logging.info("Successfully saved the training data.")
        x_test.to_csv("x_test.csv", index=False)
        y_test.to_csv("y_test.csv", index=False)
        logging.info("Successfully saved the testing data.")
    except Exception as ex:
        logging.info(f"The error raise When save the train and test data {ex} ")
    logging.info("Successfully Completed----------------------------------------------------------")
except Exception as e:
    logging.info(f"The Model selection NOT Completed due to : {e}")
