==> code/3.feature_selection.ipynb <==
import pandas as pd
from sklearn.feature_selection import SelectKBest, chi2
import logging

logging.basicConfig(filename="3.feature_selection.log", level=logging.DEBUG, format='%(asctime)s %(message)s',datefmt="%Y-%m-%d %H:%M:%S")

logging.info("loading the train dataset.........")
try:
    x_train = pd.read_csv("x_train.csv")
    y_train = pd.read_csv("y_train.csv")
    logging.info("successfully load the training data .")
except Exception as ex:
    logging.info(f"The error when  load the trai data : {ex}")

logging.info("loading the test dataset.........")
try:
    x_test = pd.read_csv("x_test.csv")
    y_test = pd.read_csv("y_test.csv")
    logging.info("successfully load the test data .")
except Exception as ex:
    logging.info(f"The error when  load the test data : {ex}")

#  selectKbest() :used to reduce the number of features.

def feature_selection(x,y):
   
    best = SelectKBest(chi2, k=4)
    best.fit_transform(x,y)
    col= x.columns
    f = best.get_support()
    df1 = col[f]
    logging.info(f"best features are :{df1}")
    return df1

logging.info("4 best features are............")
features = feature_selection(x_train, y_train)
logging.info(features)
