==> code/4.model_selction.ipynb <==
import pandas as pd
import logging
import warnings
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from xgboost import XGBClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import accuracy_score, recall_score
warnings.filterwarnings('ignore')
logging.basicConfig(filename="4.Model_selection.log", level=logging.INFO, format='%(asctime)s %(message)s',datefmt="%Y-%m-%d %H:%M:%S")
try:
    logging.info("The Model_selection process is Startes")
    try:
        x_train =  pd.read_csv("x_train.csv")
        y_train =  pd.read_csv("y_train.csv")
        logging.info("Sucessfully loading the train dataset========================")
    except Exception as e:
        logging.info(f"The error raise when loading the train dataset:{e}========================")
    try:  
        x_test =  pd.read_csv("x_test.csv")
        y_test =  pd.read_csv("y_test.csv")
        logging.info("Sucessfully loading the test dataset========================")
    except Exception as e:
        logging.info(f"The error raise when loading the test dataset:{e}========================")
    try:
        pipe1 = Pipeline([("minmax_scalar", MinMaxScaler()), ("logistic_regression", LogisticRegression())])
        pipe2 = Pipeline([("minmax_scalar", MinMaxScaler()), ("svm", SVC())])
        pipe3 = Pipeline([("minmax_scalar", MinMaxScaler()), ("KNN", KNeighborsClassifier())])
        pipe4 = Pipeline([("minmax_scalar", MinMaxScaler()), ("XGboost", XGBClassifier())])
        pipe5 = Pipeline([("minmax_scalar", MinMaxScaler()), ("decision_tree", DecisionTreeClassifier())])
        pipe6 = Pipeline([("minmax_scalar", MinMaxScaler()), ("random_forest", RandomForestClassifier())])
        logging.info("Sucessfully Created the pipeline========================")
    except Exception as e:
        logging.info(f"The error raise when Created the pipeline:{e}========================")    
    pipelines = [pipe1, pipe2, pipe3, pipe4, pipe5, pipe6]
    try:
        for pipe in pipelines:
            pipe.fit(x_train, y_train)
            logging.info("sucessfully train to the model")
    except Exception as e:
        logging.info(f"The error raise when train to the model:{e}========================")  
    try:
        y_pd1 = pipe1.predict(x_test)
        y_pd2 = pipe2.predict(x_test)
        y_pd3 = pipe3.predict(x_test)
        y_pd4 = pipe4.predict(x_test)
        y_pd5 = pipe5.predict(x_test)
        y_pd6 = pipe6.predict(x_test)
        logging.info("Sucessfully predicted the Y_prediction")
    except Exception as e:
        logging.info(f"The error raise when predicted the Y_prediction:{e}========================")
    try:
        logging.info("the performance metrics of all the model.......")
        logging.info(f"The Accuracy of Logistic_Regression {(round(accuracy_score(y_test,y_pd2)*100,2))}" )
        logging.info(f"The Recall of Logistic_Regression {(round(recall_score(y_test,y_pd2),2))}")
        logging.info("--------------------------------------------------------------------")
        logging.info(f"The Accuracy of SVC {(round(accuracy_score(y_test,y_pd2)*100,2))}" )
        logging.info(f"The Recall of SVC {(round(recall_score(y_test,y_pd2),2))}" )
        logging.info("--------------------------------------------------------------------")
        logging.info(f"The Accuracy of KNN {(round(accuracy_score(y_test,y_pd3)*100, 2))}")
        logging.info(f"The Recall of KNN {(round(recall_score(y_test,y_pd3),2))}")
        logging.info("--------------------------------------------------------------------")
        logging.info(f"The Accuracy of xgboost {(round(accuracy_score(y_test,y_pd4)*100,2))}")
        logging.info(f"The Recall of xgboost {(round(recall_score(y_test,y_pd4),2))}")
        logging.info("--------------------------------------------------------------------")
        logging.info(f"The Accuracy of decision_tree {(round(accuracy_score(y_test,y_pd5)*100,2))}")
        logging.info(f"The Recall of decision_tree {(round(recall_score(y_test,y_pd5),2))}" )
        logging.info("--------------------------------------------------------------------")
        logging.info(f"The Accuracy of Random_forest {(round(accuracy_score(y_test,y_pd6)*100,2))}")
        logging.info(f"The Recall of Random_forest {(round(recall_score(y_test,y_pd6),2))}")
        logging.info("Sucessfullly Model Selection process is completed.")
    except Exception as e:
        logging.info(f"The error raise when during performance metrics:{e}========================")
except Exception as e:
    logging.info(f"Model selection not completed due to some  :{e}========================")
